Major progress on Saving/CSS
CSS is starting to take form. Need to mess around with the banners and such more. Perhaps get someone to come up with a layout for me. I got the color choice of a blind man. Saving and displaying are getting better. Added a clock. Up next: simple alert on a event reaching alert time.